Update Feature Search results triangles which span multiple blocks
When a Feature Search result spanned multiple blocks, adding the first block from the search result displayed the feature triangle, but adding the 2nd block did not.  This seemed to depend on the dataset - the gene dataset displayed if it was 2nd, but the 40k dataset did not. Recorded in #416.
axis-ticks-selected.js : selectedFeaturesEffect() : add dependency axis1d.blocks.length, so that Feature Search results triangles are displayed when viewing multiple blocks from the search results.