retain selected SNPs when another dataset is added in Genotype Table
This commit also implements reset of selected SNPs instead, via gtBlockViewEffect(), as an alternative to retain;  retain (via sampleFiltersCopyEffect) seems OK, except that view block may switch to features tab, and fresh instance of manage-genotype is not sorted;  committing to test further.
matrix-view.js : topLeftDialogUpdate() : wrap set() of topLeftDialog with later() to avoid update of attribute after it had already been used previously in the same computation, when called from updateSettings() ... afterRender().
manage-genotype.hbs : use sampleFiltersCopyEffect; add comment showing use of gtBlockViewEffect.
manage-genotype.js :
 sampleFiltersClear() : handle undefined sampleFilters or selected.
 add gtBlocksPrevious, gtBlockViewEffect(), sampleFiltersCopyEffect(), sampleFiltersCopy(), sampleFiltersCopyType(), selectedSampleRefreshDisplay(). The requirements for selectedSampleRefreshDisplay() are still being worked out.
lookupBlockWithinBlocks() : if lookupBlock is not found, search for a block with same .referenceBlock as lookupBlock, and fallback to using blocks[0] if .length because result -1 is not handled.