GUI display of error-report from upload
Closes 428.
The issue was that a typo (e.g. missing double quotes around blocks) in JSON upload displays in GUI as [object Object].
upload.js : uploadParsedCb() and uploadParsedTryCb() : replace Error( with ErrorStatus(400, .
The required change was for augmentedMessage in uploadParsedTryCb(); the other 3 Error()-s are similar so the same change was applied.
Probably required also in Dataset.upload() and handleJson(), which is used by blockFeaturesAdd - which is used from the command line not the GUI.