finish update of filtering checkboxes in Dataset explorer
These changes are part of the frontend framework upgrade #370.
manage-explorer.js : historyViewChanged() historyBlocksChanged() and showHierarchyChanged() : set the object attribute, which was previously done implicitly for the {{input}} name or checked attribute; this was changed to <input> in 004fabf5.  use Ember_set() for .controlOptions.{historyView,historyBlocks,showHierarchy} because they are tracked by a tracking context.
entry-blocks-array.js : dataFiltered() : if data is an (Array) Proxy (e.g. ManyArray), convert it to an Array, to avoid error re. mutateSortRelatedRecords.
manage-explorer.hbs : pass value=target.checked to the checkbox actions historyBlocksChanged and showHierarchyChanged