changes required by loadBlock changing to a toggle
mapview loadBlock() was changed to a toggle in 137c3662, but there are 4 calls which aim to view block, not toggle it, so pass param view=true to loadBlock() in :
  natural-search.js : nameChr2Dataset()
  manage-explorer.js : actions.loadBlock()
  manage-genotype.js : blockViewAndBrush()
  select-parent.js : actions.loadBlock()

manage-genotype.js : blockBrushDomain() : in later() check for ! block.axis1d || block.axis1d.isDestroying

entry-block-add-button.js : actions.loadBlock() : if block is toggling to un-viewed, skip setting visible / colour for Trait / Ontology.
mapview.js : loadBlock() : add optional param viewOpt, if it is provided, block.isViewed is set to that.