Genotype Table GUI changes
Part of #495 :
genotype-samples.hbs : drop tooltip which was not essential as the context was self-explanatory, and the tooltip obscured : Filter available samples by name;  for this item in #495 : The tooltip when you mouse over Filter available samples obscures part of the menu; this tooltip is not essential because it says Search/Filter and also has description within the empty text box.
mapview.hbs : hideControls : change glyphicon : expand -> cog, i.e. change the icon which enables the Genotype Table controls dialog.