Show intersection samples count for easier verification
This count is displayed in the Genotype Table controls dialog, but it is much easier to test the effect of de/selecting SNPs with this continuous display; it can be removed if not required after this test phase.
manage-genotype.hbs : show .vcfGenotypeSamplesCount in the Actions panel to help verify the functionality of intersection of multiple chromosomes' filtered samples, added in recent commit b32d03ab. Also show .brushedVCFBlocks info in hover.
app.scss : add .min-width-wide, .min-width-wide__inner for .ember-tooltip, .ember-tooltip-inner, min-width : 50em.  In this commit it is used for .vcfGenotypeSamplesCount hover with .brushedVCFBlocks info, but can be used for other wide tooltips.