enable Genotype Table Column sorting for multiple chromosomes
manage-genotype.js : vcfGenotypeSamplesDataset() : call mapSamplesToBlock() for each viewed block of the dataset, viewedBlocksOfDataset.  The samples request for the 1st block of the dataset will cache samples of dataset, so vcfGenotypeSamplesDataset() is not called for the other blocks of the dataset and they were not linked to samples.
vcf-feature.js : featureBlockColourValue() handle undefined feature.blockId.axis1d