move and restyle the Genolink link to a button
Prepended @ to the args @side and @delay of <EmberTooltip >-s which are edited / added; the remainder will be done in a separate commit.

goto-feature-list.js : change tab to spaces.
genotype-samples.hbs : move <a .genolinkSearchURL > from beside to below the <Textarea .selectedSamplesText >, and style it as a button; hide it if ! .genolinkSearchURL.
genotype-samples.js : genolinkSearchURL() : return falsey in these cases : undefined if ! .selectedSamples, 0 if there are no AGG samples selected. This enables hiding the button when there are no selected AGG samples.
haplotypes-samples.hbs : add <EmberTooltip > for {{form/select-multiple }}.
manage-genotype.hbs :
  vcfGenotypeLookup button (1st - in the dialog) : wrap with <span> to localise the sensitivity and position of the tooltip.
  Change @side of 3 <EmberTooltip >, to position the tooltip better.
  uses of .vcfGenotypeLookupButtonDisabled : prepend : Enable this button by