Show Loading message while waiting for dotplot response
pan-barlex.hbs : wrap Loading message with {{#if this.requestPromise}}
pan-barlex.js : add requestPromise. getDotPlot() : save result promise of fetchDotPlotData() as .requestPromise
axis-brush.js : simplify brushedDomain dependencies in features(), and brushedDomainRounded() and brushedDomainInt() which were based on that; the whole array .brushedDomain is replaced not mutated in brushHelper(), so .{0,1} is not required.