ensure hex string is sent in API response instead of ObjectID for blockFeaturesCount and blockFeaturesInterval
block.js : blockFeaturesCount() and blockFeaturesInterval() : use arrayFieldToHex() to convert MongoDb ObjectID to hex string representation after reading from cache and before writing to cache; this will fix existing cache entries and ensure correct format in new stored values. It seems that ObjectIDs written to cache previously converted to hex, but now JSON.stringify() of the ObjectID is stored, and when that is read back it is not fully hydrated with .toHexString() in the .__proto__, which results in that value being sent in API response instead of the hex string format.  It is likely not necessary to apply arrayFieldToHex() to a calculated result being sent in reply because that ObjectID will still have .toHexString(), and will convert to hex string when sent in API response.

mongoDB-driver-lib.js :
  drop the alias of ObjectID to ObjectId - better to stick with the library convention, rather than that of mongo shell, noted in utilities/localise-blocks.js,block-features.js
  add hexToObjectID(),  idToHex(), (the form let name = function(i) is accepted in the Web Inspector console, so is useful during initial development).
  add arrayFieldToHex().